+++ web/components/collateral-type.tsx
@@ -27,7 +27,7 @@ export const CollateralType: React.FC<CollateralTypeProps> = ({ name, url, token
         <span className="text-gray-900 font-medium">{liquidationPenalty}%</span>
       </td>
       <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-500">
-        <span className="text-gray-900 font-medium">${maximumDebt / 1000000} million</span>
+        <span className="text-gray-900 font-medium">${maximumDebt / 1000000000000} million</span>
       </td>
       <td className="px-6 py-4 text-left whitespace-nowrap text-sm text-gray-500">
         <span className="text-gray-900 font-medium">${totalDebt / 1000000}</span>